Refactor Draggable events + add examples to docs.
