+++ docs/public/examples/draggable-autoscroll.html
@@ -1,0 +1,80 @@
+<!doctype html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <title>DragDoll Examples - Draggable Box</title>
+    <meta
+      name="viewport"
+      content="user-scalable=no, width=device-width, initial-scale=1, maximum-scale=1"
+    />
+    <link rel="stylesheet" href="base.css" />
+    <style>
+      body {
+        height: 300%;
+      }
+      .drag-container {
+        position: fixed;
+        left: 10px;
+        top: 10px;
+      }
+      .draggable {
+        position: absolute;
+        left: 50%;
+        top: 50%;
+        transform: translateX(-50%) translateY(-50%);
+      }
+    </style>
+  </head>
+  <body>
+    <div class="drag-container"></div>
+    <div class="card draggable" tabindex="0"></div>
+    <script type="importmap">
+      {
+        "imports": {
+          "eventti": "../assets/eventti.js",
+          "tikki": "../assets/tikki.js",
+          "dragdoll": "../assets/dragdoll.js"
+        }
+      }
+    </script>
+    <script src="../assets/eventti.js" type="module"></script>
+    <script src="../assets/tikki.js" type="module"></script>
+    <script src="../assets/dragdoll.js" type="module"></script>
+    <script type="module">
+      import {
+        Draggable,
+        PointerSensor,
+        KeyboardMotionSensor,
+        createPointerSensorStartPredicate,
+        autoScrollPlugin,
+      } from '../assets/dragdoll.js';
+
+      const element = document.querySelector('.draggable');
+      const pointerSensor = new PointerSensor(element);
+      const keyboardSensor = new KeyboardMotionSensor();
+      const draggable = new Draggable([pointerSensor, keyboardSensor], {
+        getElements: () => [element],
+        container: document.querySelector('.drag-container'),
+        startPredicate: createPointerSensorStartPredicate(),
+      }).use(
+        autoScrollPlugin({
+          targets: [
+            {
+              element: window,
+              axis: 'y',
+              padding: { top: Infinity, bottom: Infinity },
+            },
+          ],
+        }),
+      );
+
+      draggable.on('start', () => {
+        element.classList.add('dragging');
+      });
+
+      draggable.on('end', () => {
+        element.classList.remove('dragging');
+      });
+    </script>
+  </body>
+</html>
