+++ docs/public/examples/draggable-basic.html
@@ -7,44 +7,18 @@
       name="viewport"
       content="user-scalable=no, width=device-width, initial-scale=1, maximum-scale=1"
     />
+    <link rel="stylesheet" href="base.css" />
     <style>
-      * {
-        box-sizing: border-box;
-      }
-      ::selection {
-        background: transparent;
-      }
-      ::-moz-selection {
-        background: transparent;
-      }
-      html {
-        overflow-y: scroll;
-        width: 100%;
-        height: 100%;
-        background: #222;
-      }
-      body {
-        margin: 0;
-        padding: 0;
-        width: 100%;
-        height: 300%;
-      }
       .draggable {
         position: absolute;
-        left: 0;
-        top: 0;
-        width: 100px;
-        height: 100px;
-        background-color: #ff0000;
-        transform: translateX(20px) translateY(10px);
-      }
-      .draggable:focus {
-        background-color: #fbff00;
+        left: 50%;
+        top: 50%;
+        transform: translateX(-50%) translateY(-50%);
       }
     </style>
   </head>
   <body>
-    <div class="draggable" tabindex="0"></div>
+    <div class="card draggable" tabindex="0"></div>
     <script type="importmap">
       {
         "imports": {
@@ -59,29 +33,27 @@
     <script src="../assets/dragdoll.js" type="module"></script>
     <script type="module">
       import {
+        Draggable,
         PointerSensor,
         KeyboardMotionSensor,
-        Draggable,
         createPointerSensorStartPredicate,
-        autoScrollPlugin,
       } from '../assets/dragdoll.js';
+
       const element = document.querySelector('.draggable');
       const pointerSensor = new PointerSensor(element);
       const keyboardSensor = new KeyboardMotionSensor();
       const draggable = new Draggable([pointerSensor, keyboardSensor], {
         getElements: () => [element],
         startPredicate: createPointerSensorStartPredicate(),
-      }).use(
-        autoScrollPlugin({
-          targets: [
-            {
-              element: window,
-              axis: 'y',
-              padding: { top: Infinity, bottom: Infinity },
-            },
-          ],
-        }),
-      );
+      });
+
+      draggable.on('start', () => {
+        element.classList.add('dragging');
+      });
+
+      draggable.on('end', () => {
+        element.classList.remove('dragging');
+      });
     </script>
   </body>
 </html>
